feat: complete Movento platform and Stripe integration

+++ content-service/src/main/java/com/movento/contentservice/dto/MovieDto.java
@@ -3,10 +3,12 @@ package com.movento.contentservice.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class MovieDto extends ContentDto {
     private String director;
-    private Double imdbRating;
-    private Double boxOfficeRevenue;
+    private BigDecimal imdbRating;
+    private BigDecimal boxOfficeRevenue;
 }
